+++ src/gbtoolchainid.c
@@ -16,6 +16,7 @@
 #include "sig_turborascal.h"
 #include "sig_gbforth.h"
 #include "sig_gbbasic.h"
+#include "sig_gbnim.h"
 
 #include "sig_music.h"
 
@@ -126,11 +127,10 @@ void gbtools_detect(uint8_t * p_rom_data, uint32_t rom_size, bool strict_mode) {
         check_gbbasic();
     }
 
-    // Turbo Rascal
+    // Various other toolchains
     check_turborascal();
-
-    // GB Forth
     check_gbforth();
+    check_gbnim();
 
     // Check for music drivers
     // May report multiple drivers (if found) in default output mode