+++ src/sig_gbnim.h
@@ -1,0 +1,8 @@
+
+#ifndef _SIG_GBNIM_H
+#define _SIG_GBNIM_H
+
+void check_gbnim(void);
+
+#endif // _SIG_GBNIM_H
+